Move Owner and Membership to upper level
